requirements.txt
imutils==0.5.3
matplotlib==3.3.2
numpy==1.19.2
Pillow==7.2.0
